refactor(importer): split into focused modules
Extract validation, category management, transaction building, and pending
extraction logic into separate, focused modules:

- validation.ts: Input validation and sanitization
- category-service.ts: Category management and auto-creation
- transaction-builder.ts: Transaction record building and preparation
- pending-extraction-service.ts: Pending run management

importer.ts is now a thin orchestration layer (187 lines, down from 557).

All modules include comprehensive tests. No behavior changes - pure refactor.

Closes #11
